requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2021.10.8
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.1
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==4.8.3
Markdown==3.3.7
martor==1.6.28
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
tzdata==2023.3
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
